Refactor tasks into tasks package
