+++ app/src/main/java/com/team7/cmput301/android/theirisproject/controller/RegisterController.java
@@ -9,7 +9,7 @@ package com.team7.cmput301.android.theirisproject.controller;
 import android.content.Intent;
 import android.os.Bundle;
 
-import com.team7.cmput301.android.theirisproject.RegisterUserTask;
+import com.team7.cmput301.android.theirisproject.tasks.RegisterUserTask;
 import com.team7.cmput301.android.theirisproject.model.CareProvider;
 import com.team7.cmput301.android.theirisproject.model.Patient;
 import com.team7.cmput301.android.theirisproject.model.User;
